Add more screens to navigation along with a back button
Added a PlaceBetScreen, NewsScreen, LeaderBoardScreen, and CurrentBetsScreen to the Home
stack. Along with the left header to be a back button. The hamburger menu was moved to the
right side of the screen. The back button is disabled on the home screen because navigating
back to the login page doesn't make sense.

 # Please enter the commit message for your changes. Lines starting