Changed backend functionality when placing bet
Now when user places a bet instead of storing based on gameID, the collection
stores each game based on date/time of the bet. This lets multiple bets be
made on a game without overwriting previous bets.

Also implemented further error checking to make sure that the bet amount
is valid.