Added bet checking functionality for payout
Everytime the mybets screen is accessed the page will check to see if
any activebets have been completed, and if they have will move them
into pastbets and will credit the account with the money owed if they
bet correctly.

A refresh will be implemented later to allow for this process to be
re-run in case.

Also added two more params to be stored when a bet is made, including
money to be paid out and now store the dateTime as well within the
document on firestore.