Changed leaderboard function handling + added some rendering
Made a simple UI that allows the users to choose whether they want to display
the leaderboard for who has the most correct bets or if they want to see
the leaderboard for the highest balance.
Called leaderboard function inside of default function in order to ensure
synchronicity.
Still needs to be styled.