Making final UI changes
Made final UI changes to Leaderboard, Place Bet Screen, and Scores Screen. Made changes to
the shadows on all 3 pages. Adding padding to the top of the close button in the modal on
the Scores Screen. Changed text color to white on the leaderboard displaying the number of
bet wins.